Added rat enemy in chill valley z_house
Also added an achievement for killing that rat.
In english, the humor doesn't work: it's a bookworm and
not a "rat de bibliothèque". Maybe we can adapt with a snake
depending on the language, but it's not the priority. Or saying
like in Jack Frost (the movie): it's the french branch. Don't know