Put iron ball in front of hero if overlaps: closes #21
Put iron ball in front of hero if overlaps: closes #21

Remark: The workaround of modifying the ground  below with a traversable
one (with a custom entity) is not used in the final version, due to some
issue with the engine that does not allow that to work, as already
mentioned. If the iron ball falls on the hero and they overlap, then the
iron ball is automatically moved in front of the hero.